pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pm-agent"
version = "0.1.0"
description = "A spec-driven project management agent powered by LangChain and LangGraph"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Tejas Gadre", email = "tejas@example.com"}]
license = {text = "MIT"}

dependencies = [
    # LangChain core
    "langchain>=0.3",
    "langchain-core>=0.3",
    "langgraph>=0.2",

    # LLM providers
    "langchain-google-genai>=2.0",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.2",
    "langchain-ollama>=0.2",

    # MCP
    "langchain-mcp-adapters>=0.1",

    # Ticket sources
    "PyGithub>=2.3",
    "atlassian-python-api>=3.41",

    # Document parsing
    "pypdf>=4.0",
    "python-docx>=1.1",

    # Scheduling
    "apscheduler>=3.10",

    # CLI
    "typer[all]>=0.12",

    # Config & validation
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",

    # Persistence
    "aiosqlite>=0.20",

    # Utilities
    "tenacity>=8.3",
    "rich>=13.7",
    "structlog>=24.1",

    # API
    "fastapi>=0.104",
    "uvicorn>=0.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "respx>=0.21",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
pm-agent = "pm_agent.cli.main:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "RUF"]
ignore = ["E501"]  # line-too-long, handled by formatter

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true

[tool.hatch.build.targets.wheel]
packages = ["pm_agent"]
